Tech Revenue (project in undergraduate)
This project was my very first data science project. It was a final for my Intro to Python and R course and the goal was to manually go through a csv file and store the columns, rows and values are key / value pairs within a dictionary (which is very tedious and time consuming), and then compute the same stats / simple visualizations using powerful libraries like pandas and numpy.